+++ NEWS.md
@@ -12,6 +12,7 @@
 * Small fixes in example code from the Handbook (reported by Boris Droz and Jan Specker)
 * Changed cache for filtering of feature groups to reduce cache size (#98)
 * Updated PubChem Transformations to 0.2.0
+* Replaced the `featuresSets` method for `groupFeatures` with (updated) `featuresSets` algorithm specific `groupFeatures` methods (`groupFeaturesOpenMS` etc), so that the latter can be called directly in sets workflows (#128)
 
 
 # patRoon 2.3.3